BLE: Move new gap types in ble/gap/Types.h
